refactor(connections): move serial modules into a serial package with framer-based protocol
