+++ tests/connection/prt3/test_protocol.py
@@ -18,7 +18,7 @@ import pytest
 from paradox.connections.protocols import ConnectionProtocol
 from paradox.connections.prt3.connection import PRT3SerialConnection
 from paradox.connections.prt3.protocol import PRT3Protocol
-from paradox.connections.serial_connection import SerialCommunication
+from paradox.connections.serial.connection import SerialCommunication
 
 # ---------------------------------------------------------------------------
 # Smoke / type hierarchy
